add a Win32CryptoRegistry back-end
This works for me on a Vista x64 machine, without any keyringrc.cfg file. I
could not get any of the text file back-ends to work because the 'username'
keys always conflicted with ConfigParser.  Using the registry avoids that
problem.  Note that I'm returning '2' from supported() in order to force it to
take precedence over the Win32Crypto file implementation.  Perhaps you have a
better way to do that.