Don't use the cached KWallet object if we specify custom argument to open the KWallet
When running tests in a different order than the 'usual' one,
keyring.tests.test_backend.KDEWalletCanceledTestCase.test_user_canceled was
sometime failing.
If it was run *after* the usual KWallet tests, `keyring.backend.kwallet` had a
value, which was returned directly, instead of relying upon the initialization
of the Fake objects provided by the test.

Now, we make sure not to use the cached global `keyring.backend.kwallet`
variable if we ever specify custom arguments to `keyring.backend.open_kwallet`

Globals are evil, I would have loved to remove the `keyring.backend.kwallet`
global, but it seems it's there for good reason, which are escaping my mind at
the moment.