Add ability to specify keychain for OS X backend
By default, the OS X backend uses the built-in `security` command to
store on retrieve passwords from the OS X keychain.

The `security` command, in turn, defaults to a user's
`~/Library/Keychains/login.keychain`. However, in certain settings such
as being called from `cron`, it will instead default to
`/Library/Keychains/System.keychain`, which will obviously not have the
user's passwords.

Many of the `security` subcommands have the flexibility to use a
non-default keychain as an argument. This commit gives the
default OS X Keyring backend `get_password` method the ability to accept
a keyword argument `keychain`. If given, this provides a specificy
keychain to use and permits `python-keyring` to function in unusual
environments such as `cron`, without impairing its other functions.