Make MacOS Error class inherit from KeyringError
This way it can be caught from outside without using MacOS-specific imports.

The KeyringError class specifies that it is the base class of keyring's errors. Well, it turned out it was not the base class of all of them. Keyring was still raising errors that were not subclasses of that KeyringError class. This change makes those a subclass of KeyringError too.

There is something to be said for that KeyringDenied would also have to be a subclass of KeyringLocked, and so on. However the descriptions of those errors don't match completely, and this would prevent them from being a subclass of the Error class (unless you want to get into the mess of multiple inheritance). So I think this would be a more safe approach.

With this change, if the user of the library wants to catch all errors related to the keyring, they just have to catch KeyringError. Without this change, they would have to make an import on the MacOS API details (which is giving errors for us as well when this is done on Windows). This would be difficult to discover for most developers as there is nothing about it in the documentation.